+++ test/namespace_test.js
@@ -15,7 +15,7 @@ vows.describe('Test Namespace declarations').addBatch({
         var error = markers[0];
         assert.equal(error.type, 'error', 'Type of marker');
         assert.equal(error.message.indexOf('[XQST0047]'), 0, 'Is Error [XQST0047]');
-        assert.deepEqual(error.pos, { sl: 1, sc: 0, el: 1, ec: 50 }, 'Marker Position');
+        assert.deepEqual(error.pos, { sl: 1, sc: 30, el: 1, ec: 50 }, 'Marker Position');
         var warning = markers[1];
         assert.equal(warning.type, 'warning', 'Type of marker');
     },